Add user registration and login
In this PR, the user registration and login are added. The user registration is done by the user registration controller and the user login is done by the user login controller. The user registration and login are added to the user routes. The user registration and login are tested with Postman.